moves eigendist display tests
