+++ tests/test_display.py
@@ -641,6 +641,20 @@ class TestDisplay:
                 )
             plt.close(fig)
 
+    def test_histogram_label_fail(self, einstein_img, curie_img):
+        with pytest.raises(ValueError, match="labels must have the same length"):
+            po.plot.histogram([einstein_img, curie_img], "einstein")
+
+    @pytest.mark.parametrize("batch_idx", [None, 0])
+    @pytest.mark.parametrize("channel_idx", [None, 0])
+    def test_histogram(self, einstein_img, curie_img, batch_idx, channel_idx):
+        po.plot.histogram(
+            [einstein_img, curie_img],
+            ["einstein", "curie"],
+            batch_idx=batch_idx,
+            channel_idx=channel_idx,
+        )
+
 
 def template_test_synthesis_all_plot(
     synthesis_object,
@@ -1264,3 +1278,66 @@ class TestMetamerDisplay:
                 synthesized_met_nostore, iteration, plot_penalties, axes
             )
         plt.close()
+
+
+class TestEigendistortionDisplay:
+    SMALL_DIM = 20
+
+    @pytest.mark.parametrize(
+        "model", ["frontend.OnOff.nograd", "ColorModel"], indirect=True
+    )
+    @pytest.mark.parametrize("method", ["power", "randomized_svd"])
+    @pytest.mark.parametrize("k", [2, 3])
+    @pytest.mark.filterwarnings(
+        "ignore:Randomized SVD complete!:UserWarning",
+    )
+    def test_display(self, model, einstein_img, color_img, method, k):
+        img = einstein_img if model.__class__ == po.models.OnOff else color_img
+        as_rgb = img.shape[1] == 3
+
+        img = img[..., : self.SMALL_DIM, : self.SMALL_DIM]
+        eigendist = po.Eigendistortion(img, model)
+        eigendist.synthesize(k=k, method=method, max_iter=10)
+        po.plot.eigendistortion_imshow(eigendist, eigenindex=0, as_rgb=as_rgb)
+        po.plot.eigendistortion_imshow(eigendist, eigenindex=1, as_rgb=as_rgb)
+
+        if method == "power":
+            po.plot.eigendistortion_imshow(eigendist, eigenindex=-1, as_rgb=as_rgb)
+            po.plot.eigendistortion_imshow(eigendist, eigenindex=-2, as_rgb=as_rgb)
+        elif method == "randomized_svd":  # svd only has top k not bottom k eigendists
+            with pytest.raises(ValueError, match="eigenindex must be the index"):
+                po.plot.eigendistortion_imshow(eigendist, eigenindex=-1)
+        plt.close("all")
+
+    @pytest.mark.parametrize(
+        "model", ["frontend.OnOff.nograd", "ColorModel"], indirect=True
+    )
+    @pytest.mark.parametrize("alpha", [1, [1], [1, 10], [1, 10, 10]])
+    @pytest.mark.parametrize("eigenindex", [0, [0, -1], [0, -1, 5]])
+    @pytest.mark.filterwarnings(
+        "ignore:Adding 0.5 to distortion:UserWarning",
+    )
+    def test_display_all(self, model, einstein_img, color_img, alpha, eigenindex):
+        # in this case, we're working with grayscale images
+        img = einstein_img if model.__class__ == po.models.OnOff else color_img
+        as_rgb = img.shape[1] == 3
+
+        img = img[..., : self.SMALL_DIM, : self.SMALL_DIM]
+        eigendist = po.Eigendistortion(img, model)
+        eigendist.synthesize(k=2, method="power", max_iter=10)
+        expectation = does_not_raise()
+        if isinstance(eigenindex, list):
+            if 5 in eigenindex:
+                expectation = pytest.raises(
+                    ValueError, match="eigenindex must be the index"
+                )
+            # this will get raised first
+            if isinstance(alpha, list) and len(alpha) != len(eigenindex):
+                expectation = pytest.raises(ValueError, match="If alpha is a list")
+        elif isinstance(alpha, list) and len(alpha) != 1:
+            expectation = pytest.raises(ValueError, match="If alpha is a list")
+        with expectation:
+            po.plot.eigendistortion_imshow_all(
+                eigendist, eigenindex=eigenindex, alpha=alpha, as_rgb=as_rgb
+            )
+        plt.close("all")
